metrics: Support float values
Prometheus natively handles all metric values as floats.

Becuase linkerd-metrics uses `AtomicU64` internally, float values have
not been broadly supported. Some work towards supporting float values
was added in 876ae02a, though this functionality is limited to scaling
u64 values at render-time rather than tracking floats direclty. It is
only used for process metrics currently.

This change updates the Gauge, Counter, and Histogram types to be
generic over the stored representation--a u64 or f64. Counter and Gauge
both use a common inner Value type: f64 values are encoded into the
AtomicU64 storage.

Histogram buckets are made more easily configurable by eliminating
needless newtypes and making the Infinite bound implicit.

This sets up to support new balancer metrics that report float values
(e.g., load averages).

+++ linkerd/http-metrics/src/lib.rs
@@ -32,7 +32,7 @@ impl<T: Hash + Eq, C: Hash + Eq> Report<T, requests::Metrics<C>> {
         labels: &T,
         status: Option<http::StatusCode>,
         class: &C,
-    ) -> Option<f64> {
+    ) -> Option<u64> {
         let registry = self.registry.lock();
         let requests = registry.get(labels)?.lock();
         let status = requests.by_status().get(&status)?;
